Decorate artifacts in igor.
This PR moves the knowledge of rpm and deb packages out of orca and into igor.
Artifact decoration can be feature toggled on and off at will.

The code here is not done yet, but the skeleton and most of the logic is present.

Currently this code only decorates rpm and deb packages. The design should make
it easy to add support for more package types/formats. Also I plan to add a parser
that can be configured through igor.yml. This would make it possible to add support
for new types of artifacts without touching the codebase in igor.

Please provide feedback on this both on the implementation and if this fits better
in orca.

There has been some concerns about property file handling. I think this PR keeps all
current functionality in igor intact.